fix: transform `data/lambda` file functions to server actions allowing access to node_env
