+++ data/lambda.ts
@@ -1,23 +1,5 @@
 import axios from 'axios';
 
-import { env } from '@/utils/env';
-
-type GetPresignedURLResponse = {
-  presigned_url: string;
-  file_url: string;
-};
-
-export async function getPresignedURL(file: File) {
-  const { data } = await axios.post<GetPresignedURLResponse>(
-    env.LAMBDA_FUNCTION_URL,
-    {
-      fileName: file.name,
-    },
-  );
-
-  return data;
-}
-
 export async function uploadFileToS3(
   presignedUrl: string,
   file: File,
